refactor: move components to home directory
