+++ src/components/home/Filter/Wish/WishItem/index.stories.ts
@@ -3,7 +3,7 @@ import WishItem, { WishItemProps } from '.';
 import { FILTER_WISHS } from '../../constants';
 
 const meta: Meta<WishItemProps> = {
-  title: 'components/Filter/Wish/WishItem',
+  title: 'components/home/Filter/Wish/WishItem',
   component: WishItem,
   tags: ['autodocs'],
   argTypes: {
