+++ src/components/home/FriendSelector/index.tsx
