+++ src/pages/Home.tsx
@@ -1,9 +1,9 @@
 import React from 'react';
 import Header from '@components/common/Header';
 import Footer from '@components/common/Footer';
-import FriendSelector from '@components/FriendSelector';
-import ThemeCategory from '@components/ThemeCategory';
-import Filter from '@components/Filter';
+import FriendSelector from '@components/home/FriendSelector';
+import ThemeCategory from '@components/home/ThemeCategory';
+import Filter from '@components/home/Filter';
 import Inner from '@components/common/Layout/Inner';
 
 export default function Home() {
